ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
kaggle==1.5.16
python-dotenv==1.0.0
nltk==3.8.1
gunicorn==21.2.0

